Move cache items to their own package
