+++ src/test/java/edu/jhuapl/tinkerpop/PropertyCacheTest.java
@@ -18,6 +18,8 @@ import static org.junit.Assert.*;
 
 import org.junit.Test;
 
+import edu.jhuapl.tinkerpop.cache.PropertyCache;
+
 /**
  * Test the {@link PropertyCache} object.
  */
